Delete genericNrm.xsd
This file is based on the REL-16 TS 28.623. However this is a REL-16 document only, that does not exist in REL-17.
This file was copied here from REL-16 for the sole reason to support XSD files in the REL-17-28.541. However, as it was agreed in the approved TDOC S5-205354 to remove the XSD solution set from REL-17 there is no need for this file anymore. It should be removed (it should not be copied over from REL-16).